Fixed animations
I noticed a short lag in player movement when we need to walk again
after stopping so I fixed that.

I duplicated the bullet prefab to fix the axe animation. You are free
to double check on the movements in the animation test scene before
deleting the old one or renaming what I had fixed.

Sometimes the zombie enemies get lost. It happened once when I went out
of that scene and come back to it again (I didn’t tweak anything
beforehand but it may just be a one time thing, so I am not sure).

Personally, I want the zombies to walk a little slower, but that is
entirely up to you guys if you want to fix it up.

And I added the just-axes sprites to tweak up the animation.